// 06032026/sce-front/src/app/model/control-calidad/DataPopupVerResolucion.ts
export interface DataPopupVerResolucion {
    idActa: number;
    idResolucion: number;
    numeroActa: string;
    nombreResolucion: string;
    eleccion: string;
    numeroElectores: number;
    numeroElectoresAusentes: number;
    ubigeo: string;
    archivoResolucion: any;
    numeroExpediente: string;
}